==> temperature_data/1_tidying_env_data.Rmd <==
---
title: "Tidying Environmental Data"
author: "allyson_demerlis"
date: "2025-05-07"
output: html_document
---
# load libraries
```{r}
library(tidyverse)
library(readxl)
```

# import and tidy CCC temp logger (tilt meter)
```{r}
cures_frame_tilt_temp <- read_excel("raw_data/Oct2020_Jan2022_1901102_urban-2020-10_(0)_Temperature.xlsx", sheet = "1901102_urban-2020-10_(0)_Tempe")
#logged every 15 min

#edit cures frame time
cures_frame_tilt_temp$Datetime <- as.POSIXct(cures_frame_tilt_temp$Datetime, tz="UTC")
cures_frame_tilt_temp$Datetime_EST <- with_tz(cures_frame_tilt_temp$Datetime, "America/New_York")

# Filter the data to select time points at 2-hour intervals
cures_frame_tilt_temp <- cures_frame_tilt_temp %>%
  filter(minute(Datetime_EST) == 0 & (hour(Datetime_EST) %% 2 == 0))

cures_frame_tilt_temp %>% 
  mutate(site = "CCC") %>% 
  dplyr::select(!Date:Datetime) %>% 
  dplyr::rename(Temperature = `Temperature (C)`) -> cures_frame_tilt_temp

# split date and time into two columns
cures_frame_tilt_temp %>% 
  tidyr::separate(Datetime_EST, c("Date", "Time"), sep = "\\ ", remove = FALSE) %>% 
  mutate(Date = as.Date(Date), Time = format(Datetime_EST, format = "%H:%M:%S")) -> cures_frame_tilt_temp

cures_frame_tilt_temp %>% 
  mutate(Time_period = hms(Time)) -> cures_frame_tilt_temp

max(cures_frame_tilt_temp$Datetime_EST) #"2022-02-01 12:00:00 EST"
min(cures_frame_tilt_temp$Datetime_EST) #"2020-10-12 20:00:00 EDT"
```

plot all data to visualize
```{r}
ggplot(cures_frame_tilt_temp, aes(x=Date, y=Temperature)) +
  geom_point()
```

```{r}
# trim off the first and last day because those include time points before the instrument was in the water
cures_frame_tilt_temp %>% 
  filter(Date > "2021-06-01" & Date < "2021-12-01") -> CCC_temp_2021_06_01
```


```{r}
ggplot(CCC_temp_2021_06_01, aes(x=Date, y=Temperature, color = site)) +
  geom_point(alpha = 0.7) +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))
```

mean daily temperature during entire dataset
```{r}
CCC_temp_2021_06_01 %>% 
  group_by(Date) %>% 
  summarise(mean_daily_temp = mean(Temperature), sd_daily_temp = sd(Temperature), daily_range = (max(Temperature)-min(Temperature))) %>% 
  ggplot(., aes(x=Date, y=mean_daily_temp)) +
  geom_ribbon(aes(ymin = mean_daily_temp - sd_daily_temp, ymax = mean_daily_temp + sd_daily_temp), fill = "grey70") +
  geom_line() +
  theme_bw()
```


# import and tidy KB Nursery HOBO logger
```{r}
# trim off the first and last days because those include time points before the HOBO was in the water
KB_05282021 <- read_csv("raw_data/KBNursery_2021_05_28.csv", skip = 1)
KB_05282021 %>% 
  mutate(Datetime_EST=with_tz(mdy_hms(`Date Time, GMT-04:00`, tz = "Etc/GMT+4"), tzone = "Etc/GMT+4")) %>% 
  mutate(Datetime_EST=with_tz(Datetime_EST, tzone = "America/New_York")) %>% 
  dplyr::rename(Temperature = `Temp, °C (LGR S/N: 21085862, SEN S/N: 21085862)`) %>% 
  dplyr::select(Temperature, Datetime_EST) %>% 
  filter(Datetime_EST > "2021-05-29 01:00:00" & Datetime_EST < "2021-08-11 11:00:00") %>% 
  mutate(Light = NA) -> KB_05282021

KB_08112021 <- read_csv("raw_data/KBNursery_2021_08_11.csv", skip = 1)
KB_08112021 %>% 
  mutate(Datetime_EST=with_tz(mdy_hm(`Date Time, GMT-04:00`, tz = "Etc/GMT+4"), tzone = "Etc/GMT+4")) %>% 
  mutate(Datetime_EST=with_tz(Datetime_EST, tzone = "America/New_York")) %>% 
  dplyr::rename(Temperature = `Temp, °C (LGR S/N: 20831178, SEN S/N: 20831178)`) %>% 
  dplyr::rename(Light = `Intensity, lum/ft² (LGR S/N: 20831178, SEN S/N: 20831178)`) %>% 
  filter(Datetime_EST < "2021-12-07 17:00:00") %>% 
  dplyr::select(Temperature, Light, Datetime_EST) -> KB_08112021

KB_120821 <- read_csv("raw_data/KBNursery_2021_12_08.csv", skip = 1)
KB_120821%>% 
  mutate(Datetime_EST=with_tz(mdy_hm(`Date Time, GMT-05:00`, tz = "Etc/GMT+5"), tzone = "Etc/GMT+5")) %>% 
  mutate(Datetime_EST=with_tz(Datetime_EST, tzone = "America/New_York")) %>% 
  dplyr::rename(Temperature = `Temp, °C (LGR S/N: 21236769, SEN S/N: 21236769)`) %>% 
  dplyr::rename(Light = `Intensity, lum/ft² (LGR S/N: 21236769, SEN S/N: 21236769)`) %>% 
  filter(Datetime_EST > "2021-12-10 04:00:00" & Datetime_EST < "2021-12-31 00:00:00") %>% 
  dplyr::select(Temperature, Light, Datetime_EST) -> KB_120821

rbind(KB_05282021, KB_08112021, KB_120821) -> all_KB_data

all_KB_data %>% 
  mutate(site = "Nursery") -> all_KB_data

all_KB_data$Temperature <- as.numeric(all_KB_data$Temperature)

all_KB_data %>% 
  tidyr::separate(Datetime_EST, c("Date", "Time"), sep = "\\ ", remove = FALSE) %>% 
  mutate(Date = as.Date(Date), Time = format(Datetime_EST, format = "%H:%M:%S")) -> all_KB_data

all_KB_data %>% 
  mutate(Time_period = hms(Time)) -> all_KB_data
```

plot all raw data 
```{r}
ggplot(all_KB_data, aes(x=Datetime_EST, y=Temperature, color = site)) +
  geom_point() +
  theme_bw() +
  scale_x_datetime(date_labels = "%b %Y",  # Customize the date format
                     date_breaks = "1 month") +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) 
```


mean daily temperature during entire dataset
```{r}
all_KB_data %>% 
  group_by(Date) %>% 
  summarise(mean_daily_temp = mean(Temperature), sd_daily_temp = sd(Temperature), daily_range = (max(Temperature)-min(Temperature))) %>% 
  ggplot(., aes(x=Date, y=mean_daily_temp)) +
  geom_ribbon(aes(ymin = mean_daily_temp - sd_daily_temp, ymax = mean_daily_temp + sd_daily_temp), fill = "grey70") +
  geom_line() +
  theme_bw()
```


```{r}
all_KB_data %>% 
  mutate(year = year(Date), month = month(Date), day = day(Date)) %>% 
  filter(Date > "2021-06-01") %>% 
  mutate(Date = as.factor(Date)) %>% 
  ggplot(., aes(x=Time_period, y=Temperature, color = Date)) +
  geom_point() +
  geom_smooth(se=F) +
  facet_wrap(~month) +
  scale_x_time(labels = scales::time_format("%H:%M")) +
  theme_bw() +
  theme(legend.position = "none")
```



# join both datasets and plot temps 
```{r}
all_KB_data %>% 
  full_join(., CCC_temp_2021_06_01) %>% 
  group_by(Date, site) %>% 
  summarise(mean_daily_temp = mean(Temperature), sd_daily_temp = sd(Temperature), daily_range = (max(Temperature)-min(Temperature))) %>% 
  filter(Date < "2021-11-03" & Date > "2021-06-20") %>% 
  ggplot(., aes(x=Date, y=mean_daily_temp, color = site)) +
  geom_ribbon(aes(ymin = mean_daily_temp - sd_daily_temp, ymax = mean_daily_temp + sd_daily_temp, fill = site), alpha = 0.3, color = NA) +
  geom_line() +
  theme_bw(base_size = 14) +
  labs(x="Date", y = "Temp (ºC)") +
  scale_color_manual(values = c("CCC" = "orange", "Nursery" = "darkblue")) +
  scale_fill_manual(values = c("CCC" = "orange", "Nursery" = "darkblue")) +
  scale_x_date(date_labels = "%b %Y",  # Customize the date format
                     date_breaks = "1 month") +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))
ggsave("CCCvsnursery_temp_v2.pdf", width=8, height = 7)
```

# save tidied dataset
```{r}
all_KB_data %>% 
  full_join(., CCC_temp_2021_06_01) %>% 
  filter(Date < "2021-11-05" & Date > "2021-06-01") %>% 
  mutate(year = year(Date), month = month(Date), day = day(Date)) %>% 
  as.data.frame() %>% 
  write_csv(., "tidy_KB_CCC_data.csv")
```
