==> temperature_data/2_thermalvariability.Rmd <==
---
title: "2_thermalvariability"
author: "allyson_demerlis"
date: "2024-10-17"
output: html_document
---

#load libraries and data

```{r}
library(tidyverse)

tidy_KB_CCC_data <- read_csv("results_csv/tidy_KB_CCC_data.csv")
```

# diel variability for experiment

```{r}
tidy_KB_CCC_data %>% 
  filter(Date < "2021-11-01" & Date > "2021-06-23") %>% 
  ggplot(., aes(x=Date, y=Temperature, color = site)) +
  geom_line() +
  theme_bw() +
  scale_color_manual(values = c("CCC" = "orange", "Nursery" = "darkblue")) +
  scale_x_date(date_labels = "%b %Y",  # Customize the date format
                     date_breaks = "1 month") +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  labs(x="Month Year", y = "Temperature")
#ggsave("temps_experimentalperiod.pdf")
```

looking at just a couple of days
```{r}
tidy_KB_CCC_data %>% 
  filter(Date <= "2021-10-30" & Date >= "2021-10-28") %>% 
  ggplot(., aes(x=Datetime_EST, y=Temperature, color = site)) +
  geom_point() +
  geom_line() +
  theme_bw() +
  scale_color_manual(values = c("CCC" = "orange", "Nursery" = "darkblue")) + 
  labs(x="Date", y = "Temperature")  +
  scale_x_datetime(date_breaks = "3 hours", date_labels = "%H:%M") +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))
```


daily range (thermal variability)
```{r}
tidy_KB_CCC_data %>% 
  group_by(Date,site) %>% 
  summarise(mean_daily_temp = mean(Temperature), sd_daily_temp = sd(Temperature), daily_range = (max(Temperature)-min(Temperature))) %>% 
  ggplot(., aes(x=Date, y=daily_range, color = site)) +
  geom_line() +
  scale_color_manual(values = c("CCC" = "orange", "Nursery" = "darkblue")) +
  theme_bw() 
```


```{r}
tidy_KB_CCC_data %>% 
  filter(site == "CCC") %>% 
  mutate(year = year(Date), month = month(Date), day = day(Date)) %>% 
  filter(Date < "2021-12-01") %>% 
  mutate(Date = as.factor(Date)) %>% 
  ggplot(., aes(x=Time, y=Temperature, color = Date)) +
  geom_point() +
  geom_smooth(se=F) +
  facet_wrap(~month, scales = "free_x") +
  scale_x_time(labels = scales::time_format("%H:%M")) +
  theme_bw() +
  theme(legend.position = "none")
```

```{r}
tidy_KB_CCC_data %>% 
  filter(site == "Nursery") %>% 
  mutate(year = year(Date), month = month(Date), day = day(Date)) %>% 
  filter(Date < "2021-12-01") %>% 
  mutate(Date = as.factor(Date)) %>% 
  ggplot(., aes(x=Time, y=Temperature, color = Date)) +
  geom_point() +
  geom_smooth(se=F) +
  facet_wrap(~month) +
  scale_x_time(labels = scales::time_format("%H:%M")) +
  theme_bw() +
  theme(legend.position = "none") +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) 
```

```{r}
tidy_KB_CCC_data %>% 
  filter(as.Date(Datetime_EST) < "2021-12-01" & as.Date(Datetime_EST) > "2021-06-23") %>% 
  ggplot(., aes(x=Time, y=Temperature, color = site)) +
  geom_point() + 
  geom_smooth() +
  facet_wrap(~month) +
  scale_x_time(labels = scales::time_format("%H:%M"))+ 
   theme(axis.text.x = element_text(angle = 45, hjust = 1))
```

```{r}
tidy_KB_CCC_data %>% 
  filter(as.Date(Datetime_EST) < "2021-12-01" & as.Date(Datetime_EST) > "2021-06-01") %>% 
  ggplot(., aes(x=Datetime_EST, y=Temperature, color = site)) +
  geom_line() +
  facet_wrap(~month, scales = "free") +
  scale_color_manual(values = c("CCC" = "orange", "Nursery" = "darkblue")) +
  theme_bw()  +
   theme(axis.text.x = element_text(angle = 45, hjust = 1)) 
ggsave("monthly_temps.pdf", width = 10, height = 7)
```



## temperature data summary statistics

(with season)
```{r}
tidy_KB_CCC_data %>% 
  #filter(as.Date(Datetime_EST) < "2021-12-01" & as.Date(Datetime_EST) > "2021-06-01") %>% 
  group_by(as.Date(Datetime_EST), site) %>% 
  mutate(season = case_when(`as.Date(Datetime_EST)` <= "2021-05-25" ~ "Dry",
                            `as.Date(Datetime_EST)` >= "2021-05-26" & `as.Date(Datetime_EST)` <= "2021-10-15" ~ "Wet",
                            `as.Date(Datetime_EST)` >= "2021-10-16" ~ "Dry")) %>% 
  group_by(site, season) %>% 
  summarise(sample_size = n(), mean_temp = mean(Temperature), sd_temp = sd(Temperature), max_temp = max(Temperature), min_temp =min(Temperature)) %>% 
  write_csv("summary_stats_temp_data.csv")

tidy_KB_CCC_data %>% 
  filter(as.Date(Datetime_EST) < "2021-12-01") %>% 
  mutate(season = case_when(as.Date(Datetime_EST) <= "2021-05-25" ~ "Dry",
                            as.Date(Datetime_EST) >= "2021-05-26" & as.Date(Datetime_EST) <= "2021-10-15" ~ "Wet",
                            as.Date(Datetime_EST) >= "2021-10-16" ~ "Dry")) %>% 
  dplyr::select(Temperature, Datetime_EST, site, season) %>% 
  group_by(as.Date(Datetime_EST), site, season) %>% 
  mutate(mean_daily_temp = mean(Temperature), sd_daily_temp = sd(Temperature), daily_range = (max(Temperature)-min(Temperature))) %>% 
  group_by(site, season) %>% 
  summarise(average_daily_range = mean(daily_range), max_daily_range = max(daily_range))
```

(without season)
```{r}
tidy_KB_CCC_data %>% 
  #filter(as.Date(Datetime_EST) < "2021-12-01" & as.Date(Datetime_EST) > "2021-06-01") %>% 
  group_by(site) %>% 
  summarise(sample_size = n(), mean_temp = mean(Temperature), sd_temp = sd(Temperature), max_temp = max(Temperature), min_temp =min(Temperature)) %>% 
  write_csv("summary_stats_temp_data.csv")

tidy_KB_CCC_data %>% 
  dplyr::select(Temperature, Datetime_EST, site) %>% 
  group_by(as.Date(Datetime_EST), site) %>% 
  mutate(mean_daily_temp = mean(Temperature), sd_daily_temp = sd(Temperature), daily_range = (max(Temperature)-min(Temperature))) %>% 
  group_by(site) %>% 
  summarise(average_daily_range = mean(daily_range), max_daily_range = max(daily_range))
```

# daily variability for entire dataset

```{r}
tidy_KB_CCC_data %>% 
  filter(as.Date(Datetime_EST) < "2021-12-01") %>% 
  dplyr::select(Temperature, Datetime_EST, site) %>% 
  group_by(as.Date(Datetime_EST), site) %>% 
  summarise(mean_daily_temp = mean(Temperature), sd_daily_temp = sd(Temperature), daily_range = (max(Temperature)-min(Temperature))) %>% 
  ggplot(., aes(x=`as.Date(Datetime_EST)`, y=daily_range, color = site)) +
  geom_line() +
  theme_bw() +
  scale_color_manual(values = c("CCC" = "orange", "Nursery" = "darkblue")) +
  scale_x_date(date_labels = "%b",  # Customize the date format
                     date_breaks = "1 month") +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  labs(x="Month", y = "Daily Variability (ºC/day)")
```

# mean daily variability by season
```{r}
#ian's code for season: phDataRaw$rDateTime[i]>"2020-5-26" && phDataRaw$rDateTime[i]<"2020-10-15"){phDataRaw$precipitation[i] = "wet"}

tidy_KB_CCC_data %>% 
  filter(as.Date(Datetime_EST) < "2021-12-01") %>% 
  group_by(as.Date(Datetime_EST), site) %>% 
  summarise(daily_variability = (max(Temperature)-min(Temperature))) %>% 
  mutate(season = case_when(`as.Date(Datetime_EST)` <= "2021-05-25" ~ "Dry",
                            `as.Date(Datetime_EST)` >= "2021-05-26" & `as.Date(Datetime_EST)` <= "2021-10-15" ~ "Wet",
                            `as.Date(Datetime_EST)` >= "2021-10-16" ~ "Dry")) %>% 
  ggplot(., aes(x=site, y=daily_variability, fill = site)) +
  geom_violin() +
  theme_classic() +
  scale_fill_manual(values = c("CCC" = "orange", "Nursery" = "darkblue")) +
  facet_wrap(~season) + 
  stat_summary(fun.y=mean, geom="point", shape=18,
                 size=3, color="black")
#ggsave("averagedailyvariabilitybyseason.pdf", width = 8, height = 6)
```


## stats

```{r}
tidy_KB_CCC_data %>% 
  filter(as.Date(Datetime_EST) < "2021-12-01" & as.Date(Datetime_EST) > "2020-12-01") %>% 
  group_by(as.Date(Datetime_EST), site) %>% 
  summarise(daily_variability = (max(Temperature)-min(Temperature))) %>% 
  mutate(season = case_when(`as.Date(Datetime_EST)` <= "2021-05-25" ~ "Dry",
                            `as.Date(Datetime_EST)` >= "2021-05-26" & `as.Date(Datetime_EST)` <= "2021-10-15" ~ "Wet",
                            `as.Date(Datetime_EST)` >= "2021-10-16" ~ "Dry")) %>% 
  group_by(site, season) %>% 
  summarise(sample_size = n(), mean_daily_variability = mean(daily_variability), sd_variability = sd(daily_variability), range = (max(daily_variability)-min(daily_variability))) %>% 
  write_csv("summarystats_dailyvariability.csv")

tidy_KB_CCC_data %>% 
  filter(as.Date(Datetime_EST) < "2021-12-01" & as.Date(Datetime_EST) > "2020-12-01") %>% 
  group_by(as.Date(Datetime_EST), site) %>% 
  summarise(daily_variability = (max(Temperature)-min(Temperature))) %>% 
  mutate(season = case_when(`as.Date(Datetime_EST)` <= "2021-05-25" ~ "Dry",
                            `as.Date(Datetime_EST)` >= "2021-05-26" & `as.Date(Datetime_EST)` <= "2021-10-15" ~ "Wet",
                            `as.Date(Datetime_EST)` >= "2021-10-16" ~ "Dry")) %>% 
  group_by(site, season) %>% 
  mutate(mean_variability = mean(daily_variability), sd_variability = sd(daily_variability)) -> daily_variability


library(rstatix)
# Ensure 'site' and 'season' are factors
daily_variability$site <- as.factor(daily_variability$site)
daily_variability$season <- as.factor(daily_variability$season)

treat_model <- lm(daily_variability ~ site*season, data = daily_variability)
treat_model_metrics <- augment(treat_model)
plot(treat_model)

# assess normality of residuals using Shapiro-Wilk test
shapiro_test(treat_model_metrics$.resid) # very significant

# assess homogeneity of variances using Levene's Test
levene_test(.resid ~ site*season, data = treat_model_metrics) #significant

#need to do non-parametric test
#kruskal-wallis test

daily_variability$site_season <- interaction(daily_variability$site, daily_variability$season)

kruskal.test(daily_variability ~ site_season, data = daily_variability) # significant

library(dunn.test)

# Perform Dunn's test with Bonferonni correction
dunn.test(daily_variability$daily_variability, daily_variability$site_season, method = "BH") %>% 
  as.data.frame() %>% 
  write_csv("dailyvariability_siteseason_posthoc.csv")
```
