==> temperature_data/3_MMMDHW.Rmd <==
---
title: "3_MMMDHW"
author: "allyson_demerlis"
date: "2024-10-17"
output: html_document
---
# load libraries and data
```{r}
library(raster)
library(parallel)
library(ncdf4)
library(tidyverse)
library(lubridate)
library(readxl)

#Climatology_MMM <- "raw_data/ct5km_climatology_v3.1.nc" #this only has mean monthly maximum data

tidy_KB_CCC_data <- read_csv("results_csv/tidy_KB_CCC_data.csv")
```

# Extract MMM from CRW data

DO NOT RE-RUN -- just get the MMMs that are written in the notes
```{r Extract_data, cache=TRUE}
  MMM.data <- brick(Climatology_MMM) 
  #MMM.data
  
  print(MMM.data)
  
  # KB Nursery
  lat_KB <- (25.6763)
  lon_KB <- (-80.0987)
  extract.pts_KB <- cbind(lon_KB,lat_KB)
  
  #CCC 
  lat.CCC <- (25.766626)
  lon.CCC <- (-80.144812)
  extract_CCC <- cbind(lon.CCC, lat.CCC)
  
  #MMM <- raster::extract(MMM.data, extract.pts,method="bilinear")
  MMM <- raster::extract(MMM.data, extract.pts_KB,method="simple") # MMM for KB is 29.53
  MMM_CCC <- raster::extract(MMM.data, extract_CCC,method="simple") # MMM for CCC is 29.54
  
  MMM_SST_Nursery = 29.53
  MMM_SST_CCC = 29.54

  #write.csv(MMM, "KB_CRW_MMM_1985-2012.csv", row.names=FALSE)

  #write.csv(MMM_CCC, "CCC_CRW_MMM_1985-2012.csv", row.names=FALSE)
```



# calculate DHW using NOAA Coral Reef Watch MMM

code adapted from https://github.com/anampc/DHW_Uva/blob/master/A.Temperature_DHW.Rmd

historical (1981-2012) SST
```{r}
#for the DHW rollapplyr function it is critical that the dates are in ascending order and that the sites are separated

#first do nursery
tidy_KB_CCC_data %>% 
  filter(site == "Nursery") %>% 
  arrange(Datetime_EST) %>% 
  dplyr::group_by(Date) %>% 
  mutate(mean_daily_temp = mean(Temperature)) %>% 
  dplyr::select(Date, mean_daily_temp, site) %>% 
  distinct() %>% 
  mutate(HotSpot = mean_daily_temp - 29.53) %>% 
  mutate(HotSpot = ifelse(HotSpot >= 0, HotSpot, 0)) %>% 
  mutate(Stress = ifelse(HotSpot >= 1, HotSpot, 0)) %>% 
  mutate(Weeks_Stress = Stress/7) -> Long.data.DHW.nurserySST
  
Long.data.DHW.nurserySST$DHW<-zoo::rollapplyr(Long.data.DHW.nurserySST$Weeks_Stress,list(-(83:0)),sum,fill=NA)

ggplot(Long.data.DHW.nurserySST, aes(x=ymd(Date))) + 
    geom_hline(yintercept=8, linetype="dashed")+
    geom_hline(yintercept=4, linetype="dashed")+
    geom_line(aes(y = DHW)) #ok this looks better
  
# now do CCC
tidy_KB_CCC_data %>% 
  filter(site == "CCC") %>% 
  arrange(Datetime_EST) %>% 
  dplyr::group_by(Date) %>% 
  mutate(mean_daily_temp = mean(Temperature)) %>% 
  dplyr::select(Date, mean_daily_temp, site) %>% 
  distinct() %>% 
  mutate(HotSpot = mean_daily_temp - 29.53) %>%
  mutate(HotSpot = ifelse(HotSpot >= 0, HotSpot, 0)) %>% 
  mutate(Stress = ifelse(HotSpot >= 1, HotSpot, 0)) %>% 
  mutate(Weeks_Stress = Stress/7)  -> Long.data.DHW.CCCSST

Long.data.DHW.CCCSST$DHW<-zoo::rollapplyr(Long.data.DHW.CCCSST$Weeks_Stress,list(-(83:0)),sum,fill=NA)

ggplot(Long.data.DHW.CCCSST, aes(x=ymd(Date))) + 
    geom_hline(yintercept=8, linetype="dashed")+
    geom_hline(yintercept=4, linetype="dashed")+
    geom_line(aes(y = DHW)) #ok this looks better


# Plot just DHW SST data
full_join(Long.data.DHW.nurserySST, Long.data.DHW.CCCSST) %>% 
ggplot(., aes(x=Date, y = DHW, color = site)) + 
    geom_hline(yintercept=4, linetype="dashed")+
    geom_line() +
    ylab("DHW (C-week)  /  Temperature (C)") +
    xlab("Date") +
    theme_classic() +
  scale_color_manual(values = c("CCC" = "orange", "Nursery" = "darkblue")) +
  scale_x_date(limits=c(ymd("2021-08-24"), ymd("2021-12-01")), 
               date_labels = "%b %Y",  # Customize the date format
                     date_breaks = "1 month") +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))
#ggsave("DHW_linegraph.pdf")

## Highest DHW per year
#this is only the span of january to december 2021
    MaxDHW_SST<-full_join(Long.data.DHW.nurserySST, Long.data.DHW.CCCSST) %>% 
  filter(ymd(Date) >= "2021-01-01") %>% 
          group_by(site) %>%
          summarize(max = max(DHW, na.rm=TRUE))
    
# CCC = 5 DHW 
# Nursery = 4.17 DHW
```

# calculate days above bleaching threshold
```{r}
tidy_KB_CCC_data %>% 
  group_by(site, Date) %>% 
  summarise(max_temp = max(Temperature)) %>% 
  filter(max_temp > 31.5) %>% 
  group_by(site) %>% 
  summarise(count = n())
```




# plot temps with MMM and DHW
```{r}
tidy_KB_CCC_data %>%
 filter(as.Date(Datetime_EST) < "2021-12-01") %>% 
  ggplot(., aes(x=Datetime_EST, y=Temperature, color = site)) +
  geom_line() +
  theme_bw() +
  labs(x="Date", y = "Temp (ºC)") +
  scale_color_manual(values = c("CCC" = "orange", "Nursery" = "darkblue")) +
  geom_hline(yintercept = 29.54, linetype = "dashed", color = "black") +
  geom_hline(yintercept = 30.54, linetype = "dashed", color = "darkgrey") +
  scale_x_datetime(date_labels = "%b %Y",  # Customize the date format
                     date_breaks = "1 month") +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  geom_vline(xintercept=as.POSIXct(as.Date("2021-06-23")))

#ggsave("CCC_nursery_temp.pdf", width=10, height = 7)
```

# calculate MMM using in-situ data

## CCC
```{r}
cures_frame_tilt_temp <- readxl::read_excel("raw_data/Oct2020_Jan2022_1901102_urban-2020-10_(0)_Temperature.xlsx", sheet = "1901102_urban-2020-10_(0)_Tempe")

cures_frame_tilt_temp %>% 
  filter(Date >= "2021-01-01" & Date <= "2021-12-31") %>% 
  select(Date, `Temperature (C)`) %>% 
  arrange(Date) %>% 
  mutate(month = month(Date), year = year(Date)) %>% 
  group_by(month, year) %>%
  summarize(MM = mean(`Temperature (C)`, na.rm = TRUE)) %>% 
  arrange(desc(MM)) #30.52 for CCC
```


## nursery

import and tidy HOBO data
```{r}
# trim off the first and last days because those include time points before the HOBO was in the water
KB_09112020<-read_csv("raw_data/archive/KBNursery_2020_09_11.csv", skip = 1)

KB_09112020 %>% 
  mutate(Datetime_EST=with_tz(mdy_hm(`Date Time, GMT-04:00`, tz = "Etc/GMT+4"), tzone = "Etc/GMT+4")) %>% 
  mutate(Datetime_EST=with_tz(Datetime_EST, tzone = "America/New_York")) %>% 
  dplyr::rename(Temperature = `Temp, °C (LGR S/N: 20597589, SEN S/N: 20597589)`) %>% 
  dplyr::rename(Light = `Intensity, lum/ft² (LGR S/N: 20597589, SEN S/N: 20597589)`) %>% 
  dplyr::select(Temperature, Light, Datetime_EST) -> KB_09112020

KB_09112020 %>% 
  filter(Datetime_EST > "2021-01-02" & Datetime_EST < "2021-02-10") -> KB_09112020

KB_02102021 <- read_csv("raw_data/archive/KBNursery_2021_02_10.csv", skip = 1)
KB_02102021 %>% 
  mutate(Datetime_EST=with_tz(mdy_hms(`Date Time, GMT-05:00`, tz = "Etc/GMT+5"), tzone = "Etc/GMT+5")) %>% 
  mutate(Datetime_EST=with_tz(Datetime_EST, tzone = "America/New_York")) %>% 
  dplyr::rename(Temperature = `Temp, °C (LGR S/N: 20976501, SEN S/N: 20976501)`) %>% 
  dplyr::rename(Light = `Intensity, lum/ft² (LGR S/N: 20976501, SEN S/N: 20976501)`) %>% 
  dplyr::select(Temperature, Light, Datetime_EST) -> KB_02102021

KB_02102021 %>% 
  filter(Datetime_EST > "2021-02-09" & Datetime_EST < "2021-05-28") -> KB_02102021

KB_05282021 <- read_csv("raw_data/KBNursery_2021_05_28.csv", skip = 1)
KB_05282021 %>% 
  mutate(Datetime_EST=with_tz(mdy_hms(`Date Time, GMT-04:00`, tz = "Etc/GMT+4"), tzone = "Etc/GMT+4")) %>% 
  mutate(Datetime_EST=with_tz(Datetime_EST, tzone = "America/New_York")) %>% 
  dplyr::rename(Temperature = `Temp, °C (LGR S/N: 21085862, SEN S/N: 21085862)`) %>% 
  dplyr::select(Temperature, Datetime_EST) %>% 
  filter(Datetime_EST > "2021-05-29 01:00:00" & Datetime_EST < "2021-08-11 11:00:00") %>% 
  mutate(Light = NA) -> KB_05282021

KB_08112021 <- read_csv("raw_data/KBNursery_2021_08_11.csv", skip = 1)
KB_08112021 %>% 
  mutate(Datetime_EST=with_tz(mdy_hm(`Date Time, GMT-04:00`, tz = "Etc/GMT+4"), tzone = "Etc/GMT+4")) %>% 
  mutate(Datetime_EST=with_tz(Datetime_EST, tzone = "America/New_York")) %>% 
  dplyr::rename(Temperature = `Temp, °C (LGR S/N: 20831178, SEN S/N: 20831178)`) %>% 
  dplyr::rename(Light = `Intensity, lum/ft² (LGR S/N: 20831178, SEN S/N: 20831178)`) %>% 
  filter(Datetime_EST < "2021-12-07 17:00:00") %>% 
  dplyr::select(Temperature, Light, Datetime_EST) -> KB_08112021

KB_120821 <- read_csv("raw_data/KBNursery_2021_12_08.csv", skip = 1)
KB_120821%>% 
  mutate(Datetime_EST=with_tz(mdy_hm(`Date Time, GMT-05:00`, tz = "Etc/GMT+5"), tzone = "Etc/GMT+5")) %>% 
  mutate(Datetime_EST=with_tz(Datetime_EST, tzone = "America/New_York")) %>% 
  dplyr::rename(Temperature = `Temp, °C (LGR S/N: 21236769, SEN S/N: 21236769)`) %>% 
  dplyr::rename(Light = `Intensity, lum/ft² (LGR S/N: 21236769, SEN S/N: 21236769)`) %>% 
  filter(Datetime_EST > "2021-12-10 04:00:00" & Datetime_EST < "2021-12-31 00:00:00") %>% 
  dplyr::select(Temperature, Light, Datetime_EST) -> KB_120821

rbind(KB_09112020, KB_02102021, KB_05282021, KB_08112021, KB_120821) -> all_KB_data

all_KB_data$Temperature <- as.numeric(all_KB_data$Temperature)

all_KB_data %>% 
  tidyr::separate(Datetime_EST, c("Date", "Time"), sep = "\\ ", remove = FALSE) %>% 
  mutate(Date = as.Date(Date), Time = format(Datetime_EST, format = "%H:%M:%S")) -> all_KB_data

all_KB_data %>% 
  mutate(Time_period = hms(Time)) -> all_KB_data
```

visualize
```{r}
all_KB_data %>% 
  ggplot(., aes(x=Datetime_EST, y=Temperature)) +
  geom_point() +
  theme_bw() +
  scale_x_datetime(date_labels = "%b %Y",  # Customize the date format
                     date_breaks = "1 month") +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) 
```

calculate MMM
```{r}
all_KB_data %>% 
  filter(Date >= "2021-01-01" & Date <= "2021-12-31") %>% 
  select(Date, Temperature) %>% 
  arrange(Date) %>% 
  mutate(month = month(Date), year = year(Date)) %>% 
  group_by(month, year) %>%
  summarize(MM = mean(Temperature, na.rm = TRUE)) %>% 
  arrange(desc(MM)) #30.49 for nursery
```
